refactor: changing declaration file of type PopoverProps
