+++ projects/ion/src/lib/popover/component/popover.component.html
@@ -5,7 +5,7 @@
   [style.left]="left + 'px'"
   [style.position]="position"
 >
-  <div class="containter">
+  <div class="container">
     <div class="header">
       <div>
         <ion-icon
