+++ projects/ion/src/lib/popover/component/popover.component.scss
@@ -112,7 +112,7 @@ $arrow-size: 4.5px;
   }
 }
 
-.containter {
+.container {
   z-index: 1000;
 }
